Split SitemapPageViewModel and refactor SSE support
Extract SSE handling into SitemapPageViewModel+SSE.swift to fix the
SwiftLint file_length violation (was 1101 lines, limit 1000).

Relax private access on SSE-related properties and clearSyncedSliderOverrides
to allow cross-file extension access. Minor OpenHABCore housekeeping:
reorder declarations, convert block comments to doc comments, and make
SitemapEventStream.stop() non-async.
