+++ OpenHABCore/Sources/OpenHABCore/Model/OpenHABPage.swift
@@ -52,6 +52,21 @@ public class OpenHABPage: NSObject, @unchecked Sendable {
 }
 
 public extension OpenHABPage {
+    convenience init?(_ page: Components.Schemas.PageDTO?) {
+        if let page {
+            self.init(
+                pageId: page.id.orEmpty,
+                title: page.title.orEmpty,
+                link: page.link.orEmpty,
+                leaf: page.leaf ?? false,
+                widgets: page.widgets?.compactMap { OpenHABWidget($0) } ?? [],
+                icon: page.icon.orEmpty
+            )
+        } else {
+            return nil
+        }
+    }
+
     func filter(_ isIncluded: (OpenHABWidget) throws -> Bool) rethrows -> OpenHABPage {
         try OpenHABPage(
             pageId: pageId,
@@ -72,20 +87,5 @@ public extension OpenHABPage {
             }
         }
         return .notFound
-    }
-
-    convenience init?(_ page: Components.Schemas.PageDTO?) {
-        if let page {
-            self.init(
-                pageId: page.id.orEmpty,
-                title: page.title.orEmpty,
-                link: page.link.orEmpty,
-                leaf: page.leaf ?? false,
-                widgets: page.widgets?.compactMap { OpenHABWidget($0) } ?? [],
-                icon: page.icon.orEmpty
-            )
-        } else {
-            return nil
-        }
     }
 }
