+++ openHAB/Models/SitemapPageViewModel.swift
@@ -33,21 +33,21 @@ class SitemapPageViewModel: ObservableObject {
     @Published var showSearchField = false
     @Published private(set) var commandStates: [String: WidgetCommandLifecycleState] = [:]
     @Published private(set) var trackerStatus: NetworkStatus = .stopped
-    @Published private(set) var widgetUpdateVersions: [String: Int] = [:]
+    @Published var widgetUpdateVersions: [String: Int] = [:]
     @Published private(set) var rowInputs: [SitemapRowInput] = []
     @Published var navigationPath: [LinkedPageNavigation] = []
 
     let networkTracker = MainActorNetworkTracker.shared
     private var openAPIService: OpenAPIService?
     private var activeConnectionInfo: ConnectionInfo?
-    private var serverProperties: OpenHABServerProperties?
+    var serverProperties: OpenHABServerProperties?
     private var pageHandlingTask: Task<Void, Never>?
-    private var sseStreamTask: Task<Void, Never>?
+    var sseStreamTask: Task<Void, Never>?
     private var foregroundRefreshTask: Task<Void, Never>?
     private var connectionObserverTask: Task<Void, Never>?
     private var networkStatusObserverTask: Task<Void, Never>?
     private let commandDispatcher = WidgetCommandDispatcher()
-    private let sitemapEventStream = SitemapEventStream()
+    let sitemapEventStream = SitemapEventStream()
     private var defaultSitemap = ""
     private var defaultSitemapLabel = ""
     private var fallbackTitle = ""
@@ -55,9 +55,9 @@ class SitemapPageViewModel: ObservableObject {
     private var isLinkedPage = false
     private var pageNetworkStatus: NetworkStatus?
     private var pageNetworkStatusAvailable = false
-    private var sseConnected = false
-    private var sseNeedsRefreshOnReconnect = false
-    private var ssePreferred = true
+    var sseConnected = false
+    var sseNeedsRefreshOnReconnect = false
+    var ssePreferred = true
     private var activePageHandlingKey: String?
     private var activePageHandlingID: UUID?
     private var commandStateResetTasks: [String: Task<Void, Never>] = [:]
@@ -728,7 +728,7 @@ extension SitemapPageViewModel {
         }
     }
 
-    private func clearSyncedSliderOverrides(using widgets: [OpenHABWidget]) -> Int {
+    func clearSyncedSliderOverrides(using widgets: [OpenHABWidget]) -> Int {
         guard !sliderValueOverrides.isEmpty else { return 0 }
         var cleared = 0
 
@@ -971,7 +971,6 @@ extension SitemapPageViewModel {
         return true
     }
 
-    // swiftlint:disable:next async_without_await
     private func handleActiveConnection(_ connection: ConnectionInfo) async {
         let previousURL = activeConnectionInfo?.configuration.url
         let newURL = connection.configuration.url
@@ -1149,150 +1148,6 @@ extension SitemapPageViewModel {
         let command = state.commandString
         commandDispatcher.send(command, for: itemname, policy: policy, phase: phase, key: key) { [weak self] itemname, command in
             self?.sendCommand(itemname: itemname, command: command, origin: .update)
-        }
-    }
-}
-
-@MainActor
-private extension SitemapPageViewModel {
-    func shouldUseSSE() -> Bool {
-        ssePreferred && serverProperties?.hasSseSupport() == true
-    }
-
-    func startSSE(sitemap: String, pageId: String) async {
-        sseConnected = false
-        sseNeedsRefreshOnReconnect = false
-        sseStreamTask?.cancel()
-
-        logger.info("Starting sitemap SSE for \(sitemap, privacy: .public)/\(pageId, privacy: .public)")
-        await sitemapEventStream.startMonitoringNetworkIfNeeded(initialConnection: networkTracker.activeConnection)
-        let stream = await sitemapEventStream.stream(sitemap: sitemap, pageId: pageId)
-
-        sseStreamTask = Task { [weak self] in
-            guard let self else { return }
-            for await msg in stream {
-                guard !Task.isCancelled else { break }
-                handleSseOutput(msg)
-            }
-        }
-    }
-
-    func handleSseOutput(_ msg: StreamOutput<SitemapEventMessage>) {
-        switch msg {
-        case .connected:
-            logger.info("Sitemap SSE connected")
-            sseConnected = true
-            ssePreferred = true
-            isUpdating = false
-            if sseNeedsRefreshOnReconnect {
-                sseNeedsRefreshOnReconnect = false
-                logger.info("Sitemap SSE reconnected, refreshing page to recover missed events")
-                startPageHandling(
-                    forceRestart: true,
-                    reason: "sse-reconnected",
-                    preserveCurrentContent: true,
-                    recreateService: false
-                )
-            }
-        case let .disconnected(error):
-            logger.warning("Sitemap SSE disconnected: \(error?.localizedDescription ?? "nil", privacy: .public)")
-            isUpdating = false
-            guard shouldFallbackToLongPolling(after: error) else {
-                sseNeedsRefreshOnReconnect = true
-                return
-            }
-            startLongPollingFallback(error)
-        case let .event(message):
-            handleSseMessage(message)
-        }
-    }
-
-    func shouldFallbackToLongPolling(after error: (any Error)?) -> Bool {
-        guard let error else { return !sseConnected }
-        if let sseError = error as? SitemapSseError {
-            switch sseError {
-            case .unsupported:
-                return true
-            }
-        }
-        return !sseConnected
-    }
-
-    func startLongPollingFallback(_ error: (any Error)?) {
-        ssePreferred = false
-        sseStreamTask?.cancel()
-        sseStreamTask = nil
-        Task {
-            await sitemapEventStream.stop()
-        }
-
-        if let error {
-            logger.warning("SSE unavailable (\(error.localizedDescription, privacy: .public)), falling back to long polling")
-        } else {
-            logger.warning("SSE unavailable, falling back to long polling")
-        }
-        startPageHandling(
-            forceRestart: true,
-            reason: "sse-fallback",
-            preserveCurrentContent: true,
-            recreateService: false
-        )
-    }
-
-    func handleSseMessage(_ message: SitemapEventMessage) {
-        switch message {
-        case .alive:
-            return
-        case let .sitemapChanged(sitemap, pageId):
-            logger.info("SSE sitemap changed \(sitemap.orEmpty, privacy: .public)/\(pageId.orEmpty, privacy: .public)")
-            startPageHandling(
-                forceRestart: true,
-                reason: "sse-sitemap-changed",
-                preserveCurrentContent: true,
-                recreateService: false
-            )
-        case let .widget(event):
-            applySseWidgetEvent(event)
-        case let .unknown(raw):
-            logger.debug("SSE unknown event: \(raw, privacy: .public)")
-        }
-    }
-
-    func applySseWidgetEvent(_ event: OpenHABSitemapWidgetEvent) {
-        guard let currentPage else { return }
-        guard let widgetId = event.widgetId else { return }
-
-        if widgetId == pageId, let label = event.label {
-            objectWillChange.send()
-            currentPage.title = label
-            return
-        }
-
-        switch currentPage.apply(event: event) {
-        case .unchanged:
-            isUpdating = false
-        case .applied:
-            objectWillChange.send()
-            widgetUpdateVersions[widgetId, default: 0] += 1
-            _ = clearSyncedSliderOverrides(using: currentPage.widgets)
-            rebuildRowInputs()
-            isUpdating = false
-        case .requiresPageReload:
-            logger.info("SSE widget \(widgetId, privacy: .public) requires full sitemap reload")
-            startPageHandling(
-                forceRestart: true,
-                reason: "sse-widget-reload-required",
-                preserveCurrentContent: true,
-                recreateService: false
-            )
-        case .notFound:
-            logger.info("SSE widget \(widgetId, privacy: .public) not found, reloading sitemap")
-            startPageHandling(
-                forceRestart: true,
-                reason: "sse-widget-missing",
-                preserveCurrentContent: true,
-                recreateService: false
-            )
         }
     }
 }
